more analysis consolidation
