+++ src/main/scala/analysis/VSA.scala
@@ -29,15 +29,13 @@ case class LiteralValue(expr: BitVecLiteral) extends Value {
   override def toString: String = "Literal(" + expr + ")"
 }
 
-trait MemoryRegionValueSetAnalysis:
-
-  val cfg: ProgramCfg
-  val globals: Map[BigInt, String]
-  val externalFunctions: Map[BigInt, String]
-  val globalOffsets: Map[BigInt, BigInt]
-  val subroutines: Map[BigInt, String]
-  val mmm: MemoryModelMap
-  val constantProp: Map[CfgNode, Map[Variable, FlatElement[BitVecLiteral]]]
+trait ValueSetAnalysis(cfg: ProgramCfg,
+                        globals: Map[BigInt, String],
+                        externalFunctions: Map[BigInt, String],
+                        globalOffsets: Map[BigInt, BigInt],
+                        subroutines: Map[BigInt, String],
+                        mmm: MemoryModelMap,
+                        constantProp: Map[CfgNode, Map[Variable, FlatElement[BitVecLiteral]]]) {
 
   val powersetLattice: PowersetLattice[Value] = PowersetLattice()
 
@@ -135,11 +133,11 @@ trait MemoryRegionValueSetAnalysis:
                 evaluateExpression(storeValue, constantProp(n)) match
                   case Some(bitVecLiteral: BitVecLiteral) =>
                     s + (r -> Set(getValueType(bitVecLiteral)))
-                    /*
-                  // TODO constant prop returned BOT OR TOP. Merge regions because RHS could be a memory loaded address
-                  case variable: Variable =>
-                    s + (r -> s(variable))
-                    */
+                  /*
+                // TODO constant prop returned BOT OR TOP. Merge regions because RHS could be a memory loaded address
+                case variable: Variable =>
+                  s + (r -> s(variable))
+                  */
                   case None =>
                     storeValue.match {
                       case v: Variable =>
@@ -159,38 +157,24 @@ trait MemoryRegionValueSetAnalysis:
 
   /** Transfer function for state lattice elements.
     */
-  def localTransfer(n: CfgNode, s: lattice.sublattice.Element): lattice.sublattice.Element =
-    n match {
-      case entry: CfgFunctionEntryNode =>
-        mmm.pushContext(entry.data.name)
-        s
-      case _: CfgFunctionExitNode =>
-        mmm.popContext()
-        s
-      case cmd: CfgCommandNode =>
-        eval(cmd.data, s, n)
-      case _ => s // ignore other kinds of nodes
-    }
-
-/** Base class for memory region analysis (non-lifted) lattice.
-  */
-abstract class ValueSetAnalysis(
-    val cfg: ProgramCfg,
-    val globals: Map[BigInt, String],
-    val externalFunctions: Map[BigInt, String],
-    val globalOffsets: Map[BigInt, BigInt],
-    val subroutines: Map[BigInt, String],
-    val mmm: MemoryModelMap,
-    val constantProp: Map[CfgNode, Map[Variable, FlatElement[BitVecLiteral]]]
-) extends MemoryRegionValueSetAnalysis {
+  def localTransfer(n: CfgNode, s: lattice.sublattice.Element): lattice.sublattice.Element = n match {
+    case entry: CfgFunctionEntryNode =>
+      mmm.pushContext(entry.data.name)
+      s
+    case _: CfgFunctionExitNode =>
+      mmm.popContext()
+      s
+    case cmd: CfgCommandNode =>
+      eval(cmd.data, s, n)
+    case _ => s // ignore other kinds of nodes
+  }
 
   /** Transfer function for state lattice elements. (Same as `localTransfer` for simple value analysis.)
     */
   def transfer(n: CfgNode, s: lattice.sublattice.Element): lattice.sublattice.Element = localTransfer(n, s)
-
 }
 
-abstract class ValueSetAnalysisWorklistSolver(
+class ValueSetAnalysisSolver(
     cfg: ProgramCfg,
     globals: Map[BigInt, String],
     externalFunctions: Map[BigInt, String],
@@ -202,24 +186,3 @@ abstract class ValueSetAnalysisWorklistSolver(
     with SimpleMonotonicSolver[CfgNode, Map[Variable | MemoryRegion, Set[Value]], MapLattice[Variable | MemoryRegion, Set[Value], PowersetLattice[Value]]]
     with ForwardDependencies
 
-object ValueSetAnalysis:
-
-  /** Intraprocedural analysis that uses the worklist solver.
-    */
-  class WorklistSolver(
-      cfg: ProgramCfg,
-      globals: Map[BigInt, String],
-      externalFunctions: Map[BigInt, String],
-      globalOffsets: Map[BigInt, BigInt],
-      subroutines: Map[BigInt, String],
-      mmm: MemoryModelMap,
-      constantProp: Map[CfgNode, Map[Variable, FlatElement[BitVecLiteral]]]
-  ) extends ValueSetAnalysisWorklistSolver(
-        cfg,
-        globals,
-        externalFunctions,
-        globalOffsets,
-        subroutines,
-        mmm,
-        constantProp
-      )
